Safari-611.1.10/LayoutTests/media/audio-controls-timeline-in-media-document-expected.txt

Pass if the timeline of audio in a media document is properly updated during playback.

EVENT(timeupdate)
EXPECTED (timeLineValue() > '0') OK
END OF TEST
